Add files via upload
This is third exercise of the course. In this course, we first implement a logistic regression model to identify the handwritten digits from 0 to 9. Later we implement forward propagation of neural network and compare the accuracy. In this exercise we didn't train the neural network, instead we used trained value of theta